Added `AudioClock` for precise audio events - Added `rate` getter on `ActiveSound` - Converted `GameLoop` to use static members and properties, and modified internal `_process` to use delta time - Updated `Track provider` to allow load new handles when the current handle is not valid anymore - Refactored classes to use `GameLoop.time` and audio clock - Refactored `ConfigurableMod` to use new audio behavior

+++ lib/features/gameplay/presentation/painters/gameplay/base.dart
@@ -11,10 +11,10 @@ sealed class PlayfieldDrawable {
   // Here is where you store your variables
 
   /// Checks if the object is expired and will no longer be rendered
-  bool isExpired(int position) => false;
+  bool isExpired(double position) => false;
 
   /// Renders the object on the canvas
-  void paint(Canvas c, int position) {
+  void paint(Canvas c, double position) {
     if (isExpired(position)) return;
   }
 }
