+++ lib/features/gameplay/presentation/painters/gameplay/playfield.dart
@@ -5,12 +5,12 @@ class PlayfieldPainter extends CustomPainter {
   PlayfieldPainter({required this.position, required this.drawables})
     : super(repaint: Listenable.merge([position, drawables]));
 
-  final ValueNotifier<int> position;
+  final ValueNotifier<double> position;
   final ValueNotifier<List<PlayfieldDrawable>> drawables;
 
   @override
   void paint(Canvas canvas, _) {
-    final int pos = position.value;
+    final double pos = position.value;
 
     for (final drawable in drawables.value) {
       drawable.paint(canvas, pos);
